requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-jose
passlib[bcrypt]